+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Procedure when bumping the version number:
 - Error handling of `Action`s is now more complex but more powerful. In
   particular, `Action`s can now return almost arbitrary errors without nesting
   `Result`s like before.
+- Renamed `Action::Result` to `Action::Output`
 
 ## v0.1.0 - 2023-02-12